fix: working multi-node load-balancing setup
Had to resort to `network_mode: host` because of the OIDC provider.
When Ohana sends a request to the config_url, it'll respond with the
internal Docker hostname as the issuer, and redirect the user to that
URL, which is unreachable. And when I override the issuer URL to be
localhost, Ohana will complain because of the URL mismatch.

The only quick solution to this is to just make sure Ohana and the user
is on the same network, and so I'm using `network_mode: host`.