various improvements (#80)
* fix: unicode file uploads, move timeout to dial

Previously, file names were included as part of the HTTP header. This
caused issues as usually headers aren't encoded as utf8. The header is
removed, and the filename is now received as part of the form data.

There was also an issue with the `inc` http.Client. Previously, a
one-second timeout was added to speed up the failover process when a
node goes down. However, this caused issues where long-running file
uploads also got timed out.

The fix is to move the timeout to the Dial function. This way, instead
of declaring timeout based on whether the headers were received or not,
it's being done on the TCP layer, where if a TCP connection isn't
established by the given timeout, the connection breaks.

This way, any long-running connection can still go through, as long as
the initial handshake is done within the given timeout.

* FileBrowser: fix video preview overflow

* feat: serve byte-range downloads

* chonky: implement folder chain

* chonky: preview audio files

* chonky: preview text files